Move RemoveDirectivesFromFile
